+++ src/main/java/ru/hh/resumebuilderbot/BotBody.java
@@ -1,7 +1,7 @@
 package ru.hh.resumebuilderbot;
 
-import ru.hh.resumebuilderbot.MessageHandler.MessageHandler;
-import ru.hh.resumebuilderbot.QuestionGenerator.QuestionGenerator;
+import ru.hh.resumebuilderbot.message_handler.MessageHandler;
+import ru.hh.resumebuilderbot.question_generator.QuestionGenerator;
 
 import java.util.Map;
