+++ src/main/java/ru/hh/resumebuilderbot/message_handler/ShowMessageHandler.java
@@ -1,8 +1,8 @@
-package ru.hh.resumebuilderbot.MessageHandler;
+package ru.hh.resumebuilderbot.message_handler;
 
 import ru.hh.resumebuilderbot.Answer;
-import ru.hh.resumebuilderbot.QuestionGenerator.QuestionGenerator;
-import ru.hh.resumebuilderbot.QuestionGenerator.ShowAllQuestionGenerator;
+import ru.hh.resumebuilderbot.question_generator.QuestionGenerator;
+import ru.hh.resumebuilderbot.question_generator.ShowAllQuestionGenerator;
 
 public class ShowMessageHandler implements MessageHandler {
     @Override
