+++ src/main/java/ru/hh/resumebuilderbot/message_handler/UnknownMessageHandler.java
@@ -1,9 +1,9 @@
-package ru.hh.resumebuilderbot.MessageHandler;
+package ru.hh.resumebuilderbot.message_handler;
 
 
 import ru.hh.resumebuilderbot.Answer;
-import ru.hh.resumebuilderbot.QuestionGenerator.FixedQuestionGenerator;
-import ru.hh.resumebuilderbot.QuestionGenerator.QuestionGenerator;
+import ru.hh.resumebuilderbot.question_generator.FixedQuestionGenerator;
+import ru.hh.resumebuilderbot.question_generator.QuestionGenerator;
 import ru.hh.resumebuilderbot.TextsStorage;
 
 public class UnknownMessageHandler implements MessageHandler {
