+++ src/main/java/ru/hh/resumebuilderbot/question_generator/ShowAllQuestionGenerator.java
@@ -1,4 +1,4 @@
-package ru.hh.resumebuilderbot.QuestionGenerator;
+package ru.hh.resumebuilderbot.question_generator;
 
 import ru.hh.resumebuilderbot.ChatId;
 import ru.hh.resumebuilderbot.Question;
